requirements.txt
playwright
python-dotenv
pandas
openpyxl
pytest          # opcional – só se for rodar testes
